Extract Chat page markdown helpers and active-model resolver
